refactor: 파티 삭제 시, 존재하는 파티인지 검증하는 로직 추가

+++ src/main/java/wad/seoul_nolgoat/service/party/PartyService.java
@@ -103,6 +103,9 @@ public class PartyService {
     // 파티 삭제
     @Transactional
     public void deleteById(Long partyId) {
+        if (!partyRepository.existsById(partyId)) {
+            throw new ApiException(PARTY_NOT_FOUND);
+        }
         partyRepository.deleteById(partyId);
     }
